[VAS-502] feat: refactored existing tests to use methods from commons.js
